requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
antiorm==1.2.1
anyio==4.1.0
APScheduler==3.6.3
async-timeout==4.0.3
bcrypt==4.1.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.7
db==0.1.1
db-sqlite3==0.0.1
decorator==5.1.1
docopt==0.6.2
easyocr==1.7.1
exceptiongroup==1.2.0
filelock==3.12.4
Flask==3.0.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
imageio==2.31.3
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.3
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
passlib==1.7.4
Pillow==10.0.1
pipreqs==0.4.13
pyaes==1.6.1
pyasn1==0.5.1
pyclipper==1.3.0.post5
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-bidi==0.4.2
python-telegram-bot==12.4.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
schedule==0.6.0
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
soupsieve==2.5
sympy==1.12
Telethon==1.33.1
tifffile==2023.8.30
torch==2.0.1
torchvision==0.15.2
tornado==6.4
typing_extensions==4.7.1
tzlocal==5.2
urllib3==1.26.7
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
